rename libs to lib
