+++ electricitymap/contrib/lib/models/constants.py
