+++ parsers/NI.py
@@ -7,16 +7,16 @@ from logging import Logger, getLogger
 from typing import Any, Dict, List, Optional, Union
 
 import arrow
+from parsers.lib.exceptions import ParserException
 from requests import Session
 
 from electricitymap.contrib.config import ZoneKey
-from electricitymap.contrib.libs.models.events import (
+from electricitymap.contrib.lib.models.events import (
     ExchangeList,
     PriceList,
     ProductionBreakdownList,
     ProductionMix,
 )
-from parsers.lib.exceptions import ParserException
 
 TIMEZONE = "America/Managua"
